backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
tensorflow>=2.13.0
joblib>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-multipart>=0.0.6
pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.3.0
httpx>=0.25.0
